Fix bug with multiple rocks at at time
By default the prefab of the rock held a reference to the scriptable object. This wasn't an issue when only one was active at a time
but when multiple are instantiate they all share the same scriptable object. To fix this on instantiation the rock will clone all of its states